refactor: Split secp verification code into its own file
